requirements.txt
cvxopt==1.3.2
fairlearn==0.10.0
matplotlib==3.10.0
medmnist==3.0.2
numpy==2.2.3
opencv_python==4.9.0.80
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.2.2
scipy==1.15.2
tensorflow==2.18.0
torch==2.3.0
torchmetrics==1.6.1
torchvision==0.18.0
tqdm==4.65.0
gradio==3.16.2
albumentations==1.3.0
opencv-contrib-python==4.11.0.86
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.5.0
omegaconf==2.1.1
test-tube>=0.7.5
streamlit==1.12.1
einops==0.3.0
transformers==4.19.2
webdataset==0.2.5
kornia==0.6
open_clip_torch==2.0.2
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
timm==0.6.12
addict==2.4.0
yapf==0.32.0
prettytable==3.6.0
safetensors==0.2.7
basicsr==1.4.2